Pass absolute paths to test dirs to prove
This helps work around an issue with Devel::Cover. When Devel::Cover is passed
a -dir flag, it chdir's to that directory. This causes issues with test code
that does something like `use lib 't/lib'`. Doing that is a bit sloppy, but
the fix is to use FindBin, and _that_ breaks when the $0 contains a relative
path that is not findable from the current working directory.

I reported this as an issue with
Devel::Cover (https://github.com/pjcj/Devel--Cover/issues/247) but for now
this workaround at least makes it _possible_ to write tests that can be run
under Devel::Cover.