Add tzdata package to all containers
There are two reasons for this. One, it makes the environment in the container
a bit more like a "normal" system.

Two, if we attempt to install this as part of an Azure test pipeline I cannot
figure out how to make it _not_ prompt the user, no matter what env vars I set
or how I fiddle with debconf-set-selections.