Fix issues caused by macOS workspace directory changing between runs
We were installing perlbrew and our various Perls under this directory. We
would then cache this and on the next run it would be restored to an
unexpected location, causing jobs to fail because they couldn't find the @INC
dirs.

This seems to be unique to macOS. Windows & Linux always use the same
workspace directory between runs.